679f7ccc49f Prevent accesses to /content files without a GUI connection
f0a7eec5a7d * lisp/net/tramp-integration.el (shortdoc): Use `tramp--w...
fb642d9cf54 Merge branch 'emacs-30' of git.savannah.gnu.org:/srv/git/...
f1e37ae423f Avoid crashes in very large buffers with long lines
4dc9e99349d Update description of string comparison functions
3817355aed5 Add sanity check when adding Tramp functions to shortdoc
1bbb3220179 ; * lisp/erc/erc-fill.el (erc-fill-wrap-mode): Doc fix.
3f6e7e1da68 js-ts-mode: Make jsdoc's "description" block a comment (b...
1c7e00f4f65 Also condition cjk-misc font-spec on Android
b89d65df5fa ; * etc/PROBLEMS: How to install sources on SUSE Linux (b...
729b17c2026 ; * src/lread.c (Fload): Add comment.  (Bug#72433)
bfb1aee9989 Fix ':defer nil' in 'use-package'
d787817076e Document unsupported color Emoji on OpenBSD
db1f749db55 Fix regressions in Customize caused by 'widget-unselected...
f94d58008eb Avoid inserting extra space in SVG data
292fcd20099 Fix c-ts-common filling function (bug#71760)
e50d597f450 Fix missing type checks before specbind

# Conflicts:
#	lisp/progmodes/c-ts-common.el

+++ src/xdisp.c
@@ -3647,14 +3647,14 @@ get_narrowed_len (struct window *w)
 static ptrdiff_t
 get_medium_narrowing_begv (struct window *w, ptrdiff_t pos)
 {
-  int len = get_narrowed_len (w);
+  ptrdiff_t len = get_narrowed_len (w);
   return max ((pos / len - 1) * len, BEGV);
 }
 
 static ptrdiff_t
 get_medium_narrowing_zv (struct window *w, ptrdiff_t pos)
 {
-  int len = get_narrowed_len (w);
+  ptrdiff_t len = get_narrowed_len (w);
   return min ((pos / len + 1) * len, ZV);
 }
 
@@ -3704,9 +3704,9 @@ get_large_narrowing_begv (ptrdiff_t pos)
 {
   if (long_line_optimizations_region_size <= 0)
     return BEGV;
-  int len = long_line_optimizations_region_size / 2;
-  int begv = max (pos - len, BEGV);
-  int limit = long_line_optimizations_bol_search_limit;
+  ptrdiff_t len = long_line_optimizations_region_size / 2;
+  ptrdiff_t begv = max (pos - len, BEGV);
+  ptrdiff_t limit = long_line_optimizations_bol_search_limit;
   while (limit > 0)
     {
       if (begv == BEGV || FETCH_BYTE (CHAR_TO_BYTE (begv) - 1) == '\n')
@@ -3722,7 +3722,7 @@ get_large_narrowing_zv (ptrdiff_t pos)
 {
   if (long_line_optimizations_region_size <= 0)
     return ZV;
-  int len = long_line_optimizations_region_size / 2;
+  ptrdiff_t len = long_line_optimizations_region_size / 2;
   return min (pos + len, ZV);
 }
